dialect/cli separation
